Renamed "encryption" package to "security". Added a separate "encryption" package under it.
