+++ app/src/main/java/fi/methics/musap/sdk/internal/datatype/MusapLink.java
@@ -23,14 +23,14 @@ import fi.methics.musap.sdk.internal.datatype.coupling.SignaturePayload;
 import fi.methics.musap.sdk.internal.datatype.coupling.payload.PollPayload;
 import fi.methics.musap.sdk.internal.datatype.coupling.payload.UpdateDataPayload;
 import fi.methics.musap.sdk.internal.datatype.coupling.UpdateDataResponsePayload;
-import fi.methics.musap.sdk.internal.encryption.AesTransportEncryption;
-import fi.methics.musap.sdk.internal.encryption.PayloadHolder;
-import fi.methics.musap.sdk.internal.encryption.TransportEncryption;
-import fi.methics.musap.sdk.internal.encryption.keygenerator.MusapKeyGenerator;
-import fi.methics.musap.sdk.internal.encryption.keystorage.AndroidKeyStorage;
-import fi.methics.musap.sdk.internal.encryption.keystorage.KeyStorageFactory;
-import fi.methics.musap.sdk.internal.encryption.mac.HmacGenerator;
-import fi.methics.musap.sdk.internal.encryption.mac.MacGenerator;
+import fi.methics.musap.sdk.internal.security.encryption.AesTransportEncryption;
+import fi.methics.musap.sdk.internal.security.encryption.PayloadHolder;
+import fi.methics.musap.sdk.internal.security.encryption.TransportEncryption;
+import fi.methics.musap.sdk.internal.security.keygenerator.MusapKeyGenerator;
+import fi.methics.musap.sdk.internal.security.keystorage.AndroidKeyStorage;
+import fi.methics.musap.sdk.internal.security.keystorage.KeyStorageFactory;
+import fi.methics.musap.sdk.internal.security.mac.HmacGenerator;
+import fi.methics.musap.sdk.internal.security.mac.MacGenerator;
 import fi.methics.musap.sdk.internal.util.ByteaMarshaller;
 import fi.methics.musap.sdk.internal.util.MLog;
 import okhttp3.MediaType;
